Rename dependencyparser to semdep
