+++ cli/src/semgrep/dependency_aware_rule.py
@@ -5,11 +5,11 @@ from typing import List
 from typing import Tuple
 
 import semgrep.output_from_core as core
-from dependencyparser.find_lockfiles import DependencyTrie
-from dependencyparser.models import LockfileDependency
-from dependencyparser.models import PackageManagers
-from dependencyparser.package_restrictions import dependencies_range_match_any
-from dependencyparser.package_restrictions import ProjectDependsOnEntry
+from semdep.find_lockfiles import DependencyTrie
+from semdep.models import LockfileDependency
+from semdep.models import PackageManagers
+from semdep.package_restrictions import dependencies_range_match_any
+from semdep.package_restrictions import ProjectDependsOnEntry
 from semgrep.error import SemgrepError
 from semgrep.rule import Rule
 from semgrep.rule_match import RuleMatch
