+++ cli/src/semgrep/rule.py
@@ -11,7 +11,7 @@ from typing import Set
 from typing import Union
 
 import semgrep.output_from_core as core
-from dependencyparser.models import PackageManagers
+from semdep.models import PackageManagers
 from semgrep.constants import RuleSeverity
 from semgrep.error import InvalidRuleSchemaError
 from semgrep.rule_lang import EmptySpan
